+++ chat/user_connection.py
@@ -2,7 +2,7 @@ from typing import List
 
 from starlette.websockets import WebSocket
 
-from message import Message
+from chat.message import Message
 
 
 class UserConnection:
