+++ dto/chat_room_response.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ChatRoomResponse:
+    room_id: str
+    room_name: str
+    connection_count: int
+
+
+@dataclass
+class ListChatRoomsResponse:
+    chat_rooms: list[ChatRoomResponse]
